add a new strategies concern to contain excess strategy private methods
